refactor: #623 — fold legacy/commands into src/gateway
slash-commands is an inbound-message decoder (parses /foo args from chat
text, decides whether to route to admin handler or pass through to agent).
Same family as gateway/mention.ts and gateway/dedupe.ts.

- src/legacy/commands/slash-commands{,.test}.ts → src/gateway/commands{,.test}.ts
- Drop slash- prefix (gateway/ dir + SlashCommandHandler class already convey it)
- Drop Discord user ID/username comments — the type is platform-agnostic
- Drop stale path-banner header comments
- Update Discord transport import path

Deviation note posted on #623: commands fold into gateway/ rather than
top-level src/commands/, since they're a peer of mention/dedupe rather
than a standalone subsystem.
