+++ src/gateway/commands.test.ts
@@ -1,10 +1,8 @@
-// src/commands/slash-commands.test.ts — Tests for slash command handler
-
 import { describe, it, expect, vi, beforeEach } from "vitest";
-import { SlashCommandHandler, type CommandContext } from "./slash-commands.js";
-import { createMockSessionStore } from "../../test-helpers.js";
-import type { AgentRuntime } from "../../agent/runtime.js";
-import type { AgentConfig } from "../../agent/types.js";
+import { SlashCommandHandler, type CommandContext } from "./commands.js";
+import { createMockSessionStore } from "../test-helpers.js";
+import type { AgentRuntime } from "../agent/runtime.js";
+import type { AgentConfig } from "../agent/types.js";
 
 interface FakeAgent {
   id: string;
